requirements.txt

Django==2.2.6
django-extensions==2.2.6
django-filter==2.2.0
django-cors-headers==3.1.1
djangorestframework==3.10.3
future==0.18.1
django-reversion==3.0.4
django-crispy-forms==1.8.0
django-import-export==1.2.0
mysqlclient==1.4.4
django-formtools==2.1
httplib2==0.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
coreapi==2.3.3
Pillow==6.2.1